requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
arcgis=2.1.0.3=py39_10
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h7f8727e_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=22.1.0=py39h06a4308_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py39h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.4=py39h06a4308_0
bottleneck=1.3.5=py39h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py39h27cfd23_1003
ca-certificates=2023.01.10=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.12.7=py39h06a4308_0
cffi=1.15.1=py39h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39h06a4308_0
cloudpickle=2.0.0=pyhd3eb1b0_0
conda=23.1.0=py39h06a4308_0
conda-package-handling=1.9.0=py39h5eee18b_1
contourpy=1.0.5=py39hdb19cb5_0
cryptography=38.0.1=py39h9ce1e76_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.12.0=py39h5eee18b_0
dask=2022.7.0=py_0
dask-core=2022.7.0=py39h06a4308_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
distributed=2022.7.0=py39h06a4308_0
entrypoints=0.4=py39h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.9=h6a678d5_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
fsspec=2022.11.0=py39h06a4308_0
geomet=1.0.0=py_0
glib=2.69.1=h4ff587b_1
heapdict=1.0.1=pyhd3eb1b0_0
icu=58.2=he6710b0_3
idna=3.4=py39h06a4308_0
importlib-metadata=4.11.3=py39h06a4308_0
importlib_metadata=4.11.3=hd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.16.0=py_1
ipython=7.34.0=pyh0005158_1
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.7.2=py_0
jedi=0.18.1=py39h06a4308_1
jeepney=0.7.1=pyhd3eb1b0_0
jinja2=3.1.2=py39h06a4308_0
jpeg=9e=h5eee18b_1
jsonschema=4.17.3=py39h06a4308_0
jupyter_client=6.1.12=py_0
jupyter_core=4.11.2=py39_0
jupyterlab_pygments=0.2.2=py_0
jupyterlab_widgets=3.0.3=py_1
keyring=23.4.0=py39h06a4308_0
kiwisolver=1.4.4=py39h6a678d5_0
krb5=1.19.4=h568e23c_0
ld_impl_linux-64=2.38=h1181459_1
lerc=4.0=pyh39e3cac_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libdeflate=1.8=h7f8727e_5
libedit=3.1.20221030=h5eee18b_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.3.0=h04863e7_1
libwebp-base=1.2.4=h5eee18b_1
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
locket=1.0.0=py39h06a4308_0
lxml=4.9.1=py39h1edc446_0
lz4=3.1.3=py39h27cfd23_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib=3.6.0=py39_3
matplotlib-base=3.6.0=py39_3
mistune=2.0.4=py39h06a4308_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
msgpack-python=1.0.3=py39hd09550d_0
munkres=1.1.4=py_0
nbclient=0.7.0=py_3
nbconvert=7.2.1=py_2
nbformat=5.6.1=py_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.6=py39h06a4308_0
notebook=6.4.12=py_0
numexpr=2.8.4=py39he184ba9_0
numpy=1.23.5=py39h14f4228_0
numpy-base=1.23.5=py39h31eccc5_0
oauthlib=3.2.1=py39h06a4308_0
olefile=0.46=pyhd3eb1b0_0
openssl=1.1.1t=h7f8727e_0
packaging=22.0=py39h06a4308_0
pandas=1.5.3=py39h417a72b_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_1
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.2=py39_0
pip=23.0.1=py39h06a4308_0
pluggy=1.0.0=py39h06a4308_1
prometheus_client=0.8.0=py_0
prompt_toolkit=3.0.5=py_0
psutil=5.9.0=py39h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycosat=0.6.4=py39h5eee18b_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.9.0=py_0
pyjwt=2.4.0=py_1
pykerberos=1.2.1=py39h7f8727e_3
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py39h06a4308_0
pyrsistent=0.18.0=py39heee7806_0
pyshp=2.1.3=pyhd3eb1b0_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.21.0=py39h06a4308_0
python-fastjsonschema=2.16.2=py39h06a4308_0
python-gssapi=1.8.1=py39_1
pytz=2022.7=py39h06a4308_0
pyyaml=6.0=py39h5eee18b_1
pyzmq=24.0.1=py39_1
readline=8.2=h5eee18b_0
requests=2.28.1=py39h06a4308_0
requests-gssapi=1.2.3=py_2
requests-kerberos=0.12.0=py39_0
requests-oauthlib=1.3.0=py_0
requests-toolbelt=0.9.1=pyhd3eb1b0_0
ruamel.yaml=0.17.21=py39h5eee18b_0
ruamel.yaml.clib=0.2.6=py39h5eee18b_1
secretstorage=3.3.1=py39h06a4308_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=65.5.0=py39h06a4308_0
six=1.16.0=py_0
sortedcontainers=2.4.0=pyhd3eb1b0_0
soupsieve=2.3.2.post1=py39h06a4308_0
sqlite=3.40.0=h5082296_0
stack_data=0.5.1=py_0
tblib=1.7.0=pyhd3eb1b0_0
terminado=0.17.1=py39h06a4308_0
tinycss2=1.1.1=py_0
tk=8.6.12=h1ccaba5_0
toolz=0.12.0=py39h06a4308_0
tornado=6.1=py39h27cfd23_0
tqdm=4.64.1=py39h06a4308_0
traitlets=5.5.0=py_0
typing-extensions=4.4.0=py39h06a4308_0
typing_extensions=4.4.0=py39h06a4308_0
tzdata=2022g=h04d1e81_0
ujson=5.4.0=py39h6a678d5_0
urllib3=1.26.13=py39h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39h06a4308_1
wheel=0.38.4=py39h06a4308_0
widgetsnbextension=3.6.1=py_0
xz=5.2.8=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zict=2.1.0=py39h06a4308_0
zipp=3.11.0=py39h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.4.9=haebb681_0
